+++ tests/views/test_leadorders_view.py
@@ -35,7 +35,8 @@ class TestLeadOrderView(BaseVersionedTestView):
         (models.LeadOrder, 'data/leadorders.json')
     ]
     serialize_attrs = [
-        'path', '_roles', '_actors', 'customer', 'project', 'timezone', 'assignment', 'assignments'
+        'path', '_roles', '_actors', 'customer', 'project', 'timezone', 'assignment',
+        'assignments', 'requirement_items'
     ]
     ignore_validation_fields = [
         'state_history', 'state', 'location'